requirements.txt
django
djangorestframework
django-cors-headers
psycopg2
django-cloudinary-storage
Pillow
django-rest-swagger
pyyaml
uritemplate
pymemcache
opencv-python
numpy